Extract ProjectConfig interface and create DatafileProjectConfig.
